Make CryptoCell 310 part of mbed-mbedtls on supported targets
The CMake target mbed-mbedtls-cryptocell310 was not linked by
default, resulting in a compilation error on targets that use
Mbed TLS:

    fatal error: mbedtls_device.h: No such file or directory

Other Mbed TLS drivers simply add themselve to mbed-mbedtls, so this
commits does the alignment to fix the build error.

Note: the driver and the core Mbed TLS have mutual dependency, so
they need to be one CMake target.