Trace: Change tr_debug color from gray to bright blue
The gray-on-black color code used for debug level print-out in
mbed_trace is hard to read. Bright-blue-on-black increases the
brightness of the text without over shadowing the info level
default (white-on-black), thus making it easier to read while
maintaining the original intention.

For original color set MBED_TRACE_COLOR_THEME to 1.