Remove errant `else` in `ble::impl::Gap::reset`
This commit fixes a bug that would cause an error following a shutdown and reinitialization of BLE. If, during reinitializing BLE after shutdown, the user attempted to call `ble::Gap::setAdvertisingParameters`, the bit flag in _active_sets would still be set to 1 and thus an error code of `BLE_ERROR_OPERATION_NOT_PERMITTED` would be returned.

This is because there was an errant `else` that would conditionalize the call to `_active_sets.clear()`.