Fix missing `as_entry` method in KVStoreSecurityDb
This commit introduces changes to fix a missing method, `as_entry`, and traces that were not updated properly to work with this new method.

Prior to this fix, it was not possible to use KVStoreSecurityDb as the build would fail since the `as_entry` method was not declared in the class header file.

This bug was introduced with commit 957486e0ebb57522301a82ed73ec885892e08b74 in PR ARMmbed/mbed-os#14198