lgtm: Add initial LGTM configuration file
Add a starter LGTM configuration file that tells LGTM what sort of files
we have in Mbed OS, what dependencies need installing, how to configure,
and how to build.

We tell LGTM to build Mbed OS's unit tests for the host toolchain, as a
start.

We install pip so we can install pypi's version of CMake, which is newer
than the current LGTM default version of 3.13.4.