M2354: add TRNG to device_has list
The NU_M2354 target does have a TRNG but it is missing from the device_has list
in targets.json. This commit updates this file such that DEVICE_TRNG is defined
for NU_M2354. No further additions are required to port the M2345's true random
number generator to Mbed OS.