Add missing BlockDevice initialization to test case
Prior to this, the test was somewhat broken. It seems to be the behavior of most BlockDevice implementations to "lazily" initialize themselves if they are not already initialized and another method is called on them (eg: erase). In some custom implementations, this is not the case.

While testing a custom BlockDevice implementation using this test, I encountered an issue where the `test_program_read_small_data_sizes` deinitializes the test block device. The following test, `test_unaligned_erase_blocks`, does not initialize the test block device. This put my custom block device in an inoperable state, causing the test to fail. Adding the initialization makes the test case conform to the expected behavior of all block devices, not just those that implement lazy initialization.