M2354: inherit PSA_V8_M devices
For `MCU_M2354`, the use of `device_has` in `targets.json` overrides the devices
present in the base target: `PSA_V8_M`. In particular, `DEVICE_TRNG` is missing
in `NU_M2354` builds as `TRNG` is a member of the `PSA_V8_M` `device_has` list.
As suggested by @ccli8, to ensure that `PSA_V8_M` devices are inherited by
derived targets, such as `MCU_M2354`, the correct syntax to use is
`device_has_add`.