BlueNRG-MS: Use non Arduino Uno pin names
The Arduino Uno pin aliases in `hal/PinNameAliases.h` are now
guarded by the macro `TARGET_FF_ARDUINO_UNO` (since 7c333d8) and
not globally available anymore. This results in the following
error (and similar ones) when building the driver for K64F:

    error: 'ARDUINO_UNO_SPI_MOSI' was not declared in this scope

To fix it, replace the aliases with their raw pin names. Now the
pins are identical to BlueNRG-2.