python: Don't require icetea
icetea is used for running icetea integration tests with `mbed-cli test
--icetea`, but not otherwise needed. icetea was intended for deprecation
with Mbed OS 6.0. Remove icetea from the list of default dependencies,
as the version of prettytable it depends on causes a Python dependency
conflict with other libraries' dependencies on prettytable, preventing
icetea from being installed to the same Python environment as other Mbed
Python tools.