Remove support for custom padding bytes from merge_region_list()
Supporting custom padding bytes complicates the code a lot, and appears
to be unused internally.

IntelHex will pad binary files as neccessary when writing, and supports
a custom byte for padding. However, merge_region_list() supported a
custom byte range for padding, which required a custom implementation.

The reason for removing was that this functionality complicated fixing
the expanding of the application to restrict_size, to fix #9949.