Expand application to restrict_size. Fixes #9949
Without this, the binary for an application is not expanded to
target.restrict_size. This means that if building a bootloader and an
application, the application will not be placed correctly when combined
with the bootloader binary, since the binary will be too small.

With this change, the bootloader binary is correctly expanded, and the
application is hence located at the expected address. This fixes #9949.