Greentea: Separate large FAT test
The Greentea test suite multipart_fat_filesystem contains a test case
that uses LittleFileSystem and FATFilesystem, and including this case
results in a binary too large for the flash memory of some targets.

We separate this test case into its own suite and guard this test suite
with a check for the size of the target's ROM. This is chosen over
simply guarding the individual test case inside the same suite as it
will allow the newly created suite to be skipped from CMake when the
FAT Filesystem tests are ported to CTest.

This issue was previously unidentified because the test suite was
disabled by default, requiring the `MBED_EXTENDED_TESTS` macro to be
defined, and was therefore not run in CI.