Connectivity: Mark open(NetworkStack*) as virtual if UNITTEST defined
Classes derived from InternetSocket are opened on a network stack,
either using open(NetworkStack*) or a templated function that resolves
the provided parameter to a NetworkStack* and then calls that method.

We make this method virtual when building Unit Tests, to allow it to be
Mocked with GMock. This allows unit testing of modules that depend on
InternetSocket implementations, to verify that they interact correctly
with the interface.