Implement optional `erase` method for SDBlockDevice
This commit introduces an implementation for `SDBlockDevice::erase` that is optional. It does not change the behavior of the driver without explicit action from the user.

The user can now provide an erase value in the constructor at runtime when creating an `SDBlockDevice` instance. If this value is a valid byte value (ie: [0, 255]), then calling `SDBlockDevice::erase` will program this value to the SD card using the existing `program` implementation. If this value is not a valid byte value (as it is by default, -1), then `SDBlockDevice::erase` remains a no-op.

This commit also introduces a configuration option, `sd.default-erase-value`, that lets the user configure a default erase value at compile time. If left undefined, the default value will be -1, retaining the no-op `erase` implementation. Any other value outside of [0, 255] will cause a static assertion at compile time.