Greentea: Refactor FAT Filesystem test cases to be independent
In the Greentea test suite `multipart_fat_filesystem`, all the test
cases are dependent on setup that occurs in the first case. This case
formats (with FAT) a HeapBlockDevice that is shared between cases.
Alongside this, the first case checks that sufficient heap memory is
available for the dynamic memory allocation required for the FAT format
process. This check is performed by testing if we can successfully
allocate the required number of bytes, and skipping the case if not. The
check is then propagated to the other cases by skipping those if the
HeapBlockDevice pointer is falsy, which it will be if the first test is
skipped.

We refactor the cases to be independent of one another. Now each test
case checks that sufficient memory can be allocated for FAT formatting,
performing said formatting or skipping as indicated by this check.

In doing so we replace remove some raw pointers, and replace some C
style arrays with std::array, in order to make memory management less
error prone.